+++ NEWS.md
@@ -1,3 +1,9 @@
+PENDING: Dataiter 0.37
+======================
+
+* `DataFrame`: Add placeholder attributes for columns so that tab
+  completion of columns as attributes at a shell works
+
 2022-10-16: Dataiter 0.36
 =========================